pyproject.toml
[tool.poetry]
name = "snmp-monitor"
version = "0.1.0"
description = ""
authors = ["Jasar Ayiken <jaasaar@126.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "snmp_monitor"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
asyncio = "^3.4.3"
easysnmp = "^0.2.6"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.29"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
pathlib = "^1.0.1"
aiomysql = "^0.2.0"
pymysql = "^1.1.0"
loguru = "^0.7.2"
redis = "^5.0.3"
h11 = "^0.14.0"

[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
